Rework VS Code integration API (#15)
This commit reworks a large part of the VS Code integration package by
updating the user-facing API and adding editor features.

At the heart of the user-facing API lies the "GlspVscodeConnector". This
component acts as a bridge between a diagram client and a graphical
language server implementation, both of which the extension authors
provide via the connector's options. The connector will propagate any
message sent by the server to the client and vice-versa. It intercepts
messages of certain types to add VS Code native functionality, for
example, it processes the "setMarkers" action sent by the server to
render diagnostics within VS Code.

The VS Code connector exposes methods to interact with active clients.
The "sendActionToActiveClient" method can be used to send various
GLSP-Protocol Actions to the currently active client. Extension authors
can use this action to interact with the diagram through user-triggered
VS Code commands. Extension authors can use interceptor functions in the
connector configuration to control how the VS Code integration behaves
when it receives messages.

The new package API now provides "quickstart components" that reduce the
necessary boilerplate code to get started. Quickstart components include
a helper class to launch the default GLSP language server, a default
editor provider to render webviews, and a helper class to connect a
language server to the VS Code integration.

Complete Changelog:

- Added native editor functionality
  - Clipboard support
  - SVG Export
  - Native menu contributions

- Rework API for extension authors
  - Component to bridge client and server (GlspVscodeConnector)
  - Interceptors
  - Method to interact with currently active diagram editor
  (sendActionToActiveClient)
  - Exposure of currently selected elements within the diagram editor

- Add quickstart components
  - GLSP server launcher
  - Default editor provider
  - Helper class to connect language server to VS Code integration

- Add documentation