#516 Fix invalid declaration `RequestTypeHintsAction` in diagram widget
The `RequestTypeHints` action no longer supports a `diagramType` property. If a string is passed in the constructor is is used as `RequestId` instead. 
Since we dispatch the `RequestTypeHints` action as plain action notification (i.e. not with `actionDispatcher.request()`)  it will not be handled properly and the client never receives the `SetTypeHintsAction` result. 
Fixes @eclipse-glsp/glsp/issues/#516