GLSP-1009: Provide web-extension example (#39)
* Bugfixes & documentation updates

## GLSP-894: Dispose client session on diagram close
Refactor the `panelOnDisposeListener` of the vscode connector to also send a `disposeClientSession` request (if the session has been successfully initialized before)
Fixes https://github.com/eclipse-glsp/glsp/issues/894

## GLSP-920: Also provide deselection info in VSCode selector (breaking)
Change  the `onSelectionUpdate` Event of the `glsp-vscode-connector` to provide both the `selectedElementIds` & `deselectedElementids`.
Fixes https://github.com/eclipse-glsp/glsp/issues/920

## Remove extended doc from vscode-integration README
Remove the extended  `Getting started` documentation part of the ´vscode-integration` README. This part contains a lot of
interface & type info which has not been maintained and is rather outdated.  Since the creation of this README our documentation efforts
have been shifted to the official GLSP Website and in general we try to avoid in-depth documentation in sub READMES of component repositories.
They are hard to maintain and will become outdated anyways. We try to use the website as "single source of truth" for documentation.

* GLSP-1009: Provide web-extension example

- Refactor codebase to provide entry points for both "normal" node VSCode extensions and web VSCode extensions.
(via "@eclipse-glsp/vscode-integration/browser")
- Add additonal webview-extension example that covers two usecases
  - Communication with a GLSPServer that is running in the same extension context
  - Communication with a GLSPServer that is running externally via websocket
- Add launch configuration for starting the web extension

Fixes https://github.com/eclipse-glsp/glsp/issues/1009

* Address review feedback